Conflicts:
	README

+++ src/utils/UIImage+animatedGIF.m
@@ -11,8 +11,8 @@
 
 @implementation UIImage (animatedGIF)
 
-static int delayCentisecondsForImageAtIndex(CGImageSourceRef const source, size_t const i) {
-    int delayCentiseconds = 1;
+static long delayCentisecondsForImageAtIndex(CGImageSourceRef const source, size_t const i) {
+    long delayCentiseconds = 1;
     CFDictionaryRef const properties = CGImageSourceCopyPropertiesAtIndex(source, i, NULL);
     if (properties) {
         CFDictionaryRef const gifProperties = CFDictionaryGetValue(properties, kCGImagePropertyGIFDictionary);
@@ -20,32 +20,32 @@ static int delayCentisecondsForImageAtIndex(CGImageSourceRef const source, size_
         if (gifProperties) {
             CFNumberRef const number = CFDictionaryGetValue(gifProperties, kCGImagePropertyGIFDelayTime);
             // Even though the GIF stores the delay as an integer number of centiseconds, ImageIO “helpfully” converts that to seconds for us.
-            delayCentiseconds = (int)lrint([fromCF number doubleValue] * 100);
+            delayCentiseconds = (long)lrint([fromCF number doubleValue] * 100);
         }
     }
     return delayCentiseconds;
 }
 
-static void createImagesAndDelays(CGImageSourceRef source, size_t count, CGImageRef imagesOut[count], int delayCentisecondsOut[count]) {
+static void createImagesAndDelays(CGImageSourceRef source, size_t count, CGImageRef imagesOut[count], long delayCentisecondsOut[count]) {
     for (size_t i = 0; i < count; ++i) {
         imagesOut[i] = CGImageSourceCreateImageAtIndex(source, i, NULL);
         delayCentisecondsOut[i] = delayCentisecondsForImageAtIndex(source, i);
     }
 }
 
-static int sum(size_t const count, int const *const values) {
-    int theSum = 0;
+static long sum(size_t const count, long const *const values) {
+    long theSum = 0;
     for (size_t i = 0; i < count; ++i) {
         theSum += values[i];
     }
     return theSum;
 }
 
-static int pairGCD(int a, int b) {
+static long pairGCD(long a, long b) {
     if (a < b)
         return pairGCD(b, a);
     while (true) {
-        int const r = a % b;
+        long const r = a % b;
         if (r == 0)
             return b;
         a = b;
@@ -53,8 +53,8 @@ static int pairGCD(int a, int b) {
     }
 }
 
-static int vectorGCD(size_t const count, int const *const values) {
-    int gcd = values[0];
+static long vectorGCD(size_t const count, long const *const values) {
+    long gcd = values[0];
     for (size_t i = 1; i < count; ++i) {
         // Note that after I process the first few elements of the vector, `gcd` will probably be smaller than any remaining element.  By passing the smaller value as the second argument to `pairGCD`, I avoid making it swap the arguments.
         gcd = pairGCD(values[i], gcd);
@@ -62,8 +62,8 @@ static int vectorGCD(size_t const count, int const *const values) {
     return gcd;
 }
 
-static NSArray *frameArray(size_t const count, CGImageRef const images[count], int const delayCentiseconds[count], int const totalDurationCentiseconds) {
-    int const gcd = vectorGCD(count, delayCentiseconds);
+static NSArray *frameArray(size_t const count, CGImageRef const images[count], long const delayCentiseconds[count], long const totalDurationCentiseconds) {
+    long const gcd = vectorGCD(count, delayCentiseconds);
     size_t const frameCount = totalDurationCentiseconds / gcd;
     UIImage *frames[frameCount];
     for (size_t i = 0, f = 0; i < count; ++i) {
@@ -84,9 +84,9 @@ static void releaseImages(size_t const count, CGImageRef const images[count]) {
 static UIImage *animatedImageWithAnimatedGIFImageSource(CGImageSourceRef const source) {
     size_t const count = CGImageSourceGetCount(source);
     CGImageRef images[count];
-    int delayCentiseconds[count]; // in centiseconds
+    long delayCentiseconds[count]; // in centiseconds
     createImagesAndDelays(source, count, images, delayCentiseconds);
-    int const totalDurationCentiseconds = sum(count, delayCentiseconds);
+    long const totalDurationCentiseconds = sum(count, delayCentiseconds);
     NSArray *const frames = frameArray(count, images, delayCentiseconds, totalDurationCentiseconds);
     UIImage *const animation = [UIImage animatedImageWithImages:frames duration:(NSTimeInterval)totalDurationCentiseconds / 100.0];
     releaseImages(count, images);
